Enable user specific firebase connection

+++ src/resources/firebase/exercises.tsx
@@ -10,6 +10,10 @@ export interface Exercise {
     weight: number
 }
 
+interface ExerciseDoc extends Exercise {
+    uid: string
+}
+
 export interface StoredExercise {
     id: string,
     exercise: Exercise
@@ -19,44 +23,65 @@ export const useStoredExercises = () => {
     const [exercises, setExercises] = useState(undefined as unknown as StoredExercise[]);
     
     useEffect(() => {
-        const unsubscribe = firebase
-            .firestore()
-            .collection('exercises')
-            .onSnapshot((snapshot) => {
-                const exercises: StoredExercise[] = snapshot.docs.map((doc) => {
-                    return {
-                        id: doc.id, 
-                        exercise: doc.data()
-                    } as StoredExercise
-                })
+        const user = firebase.auth().currentUser;
+        if(user !== null){
+            const unsubscribe = firebase
+                .firestore()
+                .collection('exercises')
+                .where('uid', '==', user.uid)
+                .onSnapshot((snapshot) => {
+                    const exercises: StoredExercise[] = snapshot.docs.map((doc) => {
+                        return {
+                            id: doc.id, 
+                            exercise: doc.data()
+                        } as StoredExercise
+                    })
 
-                setExercises(exercises);
-            })
-        return () => unsubscribe();
+                    setExercises(exercises);
+                })
+            return () => unsubscribe();
+        }
     }, [])
 
     return exercises; 
 }
 
 export const addExerciseToStore = (exercise: Exercise) => {
-    return firebase
-        .firestore()
-        .collection('exercises')
-        .add(exercise)
+    const user = firebase.auth().currentUser;
+    if(user !== null){
+        const exerciseDoc: ExerciseDoc = {
+            ...exercise,
+            uid: user.uid
+        }
+        return firebase
+            .firestore()
+            .collection('exercises')
+            .add(exerciseDoc)
+    }
 }
 
 export const editExerciseInStore = (storedExercise: StoredExercise) => {
-    return firebase
-        .firestore()
-        .collection("exercises")
-        .doc(storedExercise.id)
-        .update(storedExercise.exercise)
+    const user = firebase.auth().currentUser;
+    if(user !== null){
+        const exerciseDoc: ExerciseDoc = {
+            ...storedExercise.exercise,
+            uid: user.uid
+        }
+        return firebase
+            .firestore()
+            .collection("exercises")
+            .doc(storedExercise.id)
+            .update(exerciseDoc)
+    }
 }
 
 export const deleteExerciseFromStore = (storedExercise: StoredExercise) => {
-    return firebase
-        .firestore()
-        .collection("exercises")
-        .doc(storedExercise.id)
-        .delete()
+    const user = firebase.auth().currentUser;
+    if(user !== null){
+        return firebase
+            .firestore()
+            .collection("exercises")
+            .doc(storedExercise.id)
+            .delete()
+    }
 }
